Update README.md, fixing examples with bad JSON
The examples given in the README do not conform to the JSON specification. In fact, when run in the way shown, it doesn't even work. 

This commit fixes the formatting issues in the JSONs. They are mostly wrapping attribute names in quotes, and changing single quotes to double quotes.

I also made those changed to JS objects that wasn't specifically JSON, but would be nice to be formatted like JSON in case someone wants to copy the example and paste it into a .json file.